requirements.txt
pyqt6-tools
pynput
configparser